# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_TrinityKnights_Backend_internal_domain_model.CreatePaymentResponse:
    properties:
      amount:
        type: number
      expiry_date:
        type: string
      id:
        type: integer
      order_id:
        type: integer
      payment_url:
        type: string
      status:
        type: string
    type: object
...
  github_com_TrinityKnights_Backend_internal_domain_model.Error:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  github_com_TrinityKnights_Backend_internal_domain_model.EventResponse:
    properties:
      date:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      time:
        type: string
      venue_id:
        type: integer
    type: object
...
  github_com_TrinityKnights_Backend_internal_domain_model.OrderResponse:
    properties:
      date:
        type: string
      event_id:
        type: integer
      id:
        type: integer
      payment:
        $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.CreatePaymentResponse'
      quantity:
        type: integer
      tickets:
        items:
          $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.TicketResponse'
        type: array
      total_price:
        type: number
      user_id:
        type: string
    type: object
...
  github_com_TrinityKnights_Backend_internal_domain_model.PageMetadata:
    properties:
      page:
        type: integer
      size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_TrinityKnights_Backend_internal_domain_model.PaymentCallbackRequest:
    properties:
      amount:
        type: integer
      bank_code:
        type: string
      created:
        type: string
      currency:
        type: string
      description:
        type: string
      external_id:
        type: string
      failed_redirect_url:
        type: string
      id:
        type: string
      is_high:
        type: boolean
      merchant_name:
        type: string
      paid_amount:
        type: integer
      paid_at:
        type: string
      payer_email:
        type: string
      payment_channel:
        type: string
      payment_destination:
        type: string
      payment_method:
        type: string
      status:
        type: string
      success_redirect_url:
        type: string
      updated:
        type: string
      user_id:
        type: string
    type: object
  github_com_TrinityKnights_Backend_internal_domain_model.PaymentCallbackResponse:
    properties:
      status:
        type: string
    type: object
  github_com_TrinityKnights_Backend_internal_domain_model.PaymentResponse:
    properties:
      amount:
        type: number
      id:
        type: integer
      method:
        type: string
      order:
        $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.OrderResponse'
      order_id:
        type: integer
      status:
        type: string
      transaction_id:
        type: string
    type: object
...
  ? github_com_TrinityKnights_Backend_internal_domain_model.Response-array_github_com_TrinityKnights_Backend_internal_domain_model_PaymentResponse
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.PaymentResponse'
        type: array
      error:
        $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
      paging:
        $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.PageMetadata'
    type: object
...
  ? github_com_TrinityKnights_Backend_internal_domain_model.Response-github_com_TrinityKnights_Backend_internal_domain_model_CreatePaymentResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.CreatePaymentResponse'
      error:
        $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
      paging:
        $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.PageMetadata'
    type: object
...
  github_com_TrinityKnights_Backend_internal_domain_model.TicketResponse:
    properties:
      event:
        $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.EventResponse'
      event_id:
        type: integer
      id:
        type: string
      order:
        $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.OrderResponse'
      order_id:
        type: integer
      price:
        type: number
      seat_number:
        type: string
      type:
        type: string
    type: object
...
info:
  contact:
    email: jakueenak@gmail.com
    name: API Support
  description: This is an auto-generated API Docs.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Trinity Knights API
  version: "0.1"
paths:
...
  /payment:
    get:
      consumes:
      - application/json
      description: Get Payments
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      - description: Sort field
        enum:
        - id
        - order_id
        - amount
        - status
        in: query
        name: sort
        type: string
      - description: Sort order
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Response-array_github_com_TrinityKnights_Backend_internal_domain_model_PaymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
      security:
      - ApiKeyAuth: []
      summary: Get Payments
      tags:
      - Payment
  /payment/{payment_id}:
    get:
      consumes:
      - application/json
      description: Get Payment By ID
      parameters:
      - description: Payment ID
        in: path
        name: payment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Response-github_com_TrinityKnights_Backend_internal_domain_model_CreatePaymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
      security:
      - ApiKeyAuth: []
      summary: Get Payment By ID
      tags:
      - Payment
  /payment/callback:
    post:
      consumes:
      - application/json
      description: Callback Payment
      parameters:
      - description: Payment Callback Request
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.PaymentCallbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.PaymentCallbackResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
      summary: Callback Payment
      tags:
      - Payment
  /payment/search:
    get:
      consumes:
      - application/json
      description: Search Payments
      parameters:
      - description: Payment ID
        in: query
        name: id
        type: integer
      - description: Order ID
        in: query
        name: order_id
        type: integer
      - description: Amount
        in: query
        name: amount
        type: number
      - description: Status
        in: query
        name: status
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      - description: Sort field
        enum:
        - id
        - order_id
        - amount
        - status
        in: query
        name: sort
        type: string
      - description: Sort order
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Response-array_github_com_TrinityKnights_Backend_internal_domain_model_PaymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
      security:
      - ApiKeyAuth: []
      summary: Search Payments
      tags:
      - Payment
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"